Minor bug fix

+++ py/hscUtils/hscUtils.py
@@ -297,7 +297,7 @@ def zscale(img, contrast=0.25, samples=500):
     n = len(imsort)
     idx = np.arange(n)
 
-    med = imsort[n/2]
+    med = imsort[int(n/2)]
     w = 0.25
     i_lo, i_hi = int((0.5-w)*n), int((0.5+w)*n)
     # BUG: Sometimes the polyfit could fail
